refactor(test): isolate closeGlobalFetchDispatcher tests into dedicated file
The file-level vi.mock("undici") in run-main.test.ts applied to all
tests in the file, creating a hidden dependency for future argv-parsing
tests. Move the dispatcher tests to run-main-dispatcher.test.ts so the
undici mock is scoped only to the file that needs it.
